fix type error when requested doc comment does not exist
When calling getDocComment() on a reflector and the requested doc
comment does not exist the method returns false. Since we switched
to strict type hinting this can not be returned by the docComment()
function, and thus must be translated into something else. Throwing
an exception is not feasible, therefore we treat this by returning
an empty string instead.